backend/requirements.txt

fastapi
uvicorn[standard]
requests
pydantic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
google-generativeai